Implemented first version of the manual beat detection
The majority of this functionality is in the ButtonBpmGuesser class, which implements the View.OnLickListener class in order to be set the click listener of the button associated to the detection zone, and thus react only when it is needed.
Also, some small problems with the seek bar minimum simulation have been fixed, it now creates the correct beat; some comments have been added to the MainActivity to add details to the logic.
Note that currently, there is a minor (but somewhat major) bug: the beat is not played as it should be, some short tests show that it can play it with a few bpm late.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>