Rearranged main content to a fragment
Instead of having the pocket beat elements in the main activity load up in the background and the loaded fragments appear on top of them, they were moved to be used as a fragment also. Layouts were consequently reworked in order to enable the drawer navigator and to make the fragments appear below the toolbar, not under it.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>